fix: change flags type from uint8_t to uint16_t

+++ src/w3d/animation_parser.cpp
@@ -145,7 +145,7 @@ CompressedAnimChannel AnimationParser::parseCompressedChannel(ChunkReader &reade
   channel.numTimeCodes = reader.read<uint32_t>();
   channel.pivot = reader.read<uint16_t>();
   channel.vectorLen = reader.read<uint16_t>();
-  channel.flags = reader.read<uint8_t>();
+  channel.flags = reader.read<uint16_t>();
   reader.skip(4); // padding/reserved
 
   // Read time codes
